refactor(kernel): move journal transaction coordination
