+++ packages/kernel/src/composition/index.ts
@@ -2,7 +2,7 @@
 // for application composition roots without making store internals public.
 export { readContentAddressedBlob, readContentAddressedTextBlob, writeContentAddressedBlob } from "../persistence/blob/content-addressed-blob-store.ts";
 export { makeMarkdownArtifactStore } from "../persistence/markdown/markdown-artifact-store.ts";
-export { makeJournaledWriteCoordinator, makeOperationalJournaledWriteCoordinator } from "../store/write-journal-coordinator.ts";
+export { makeJournaledWriteCoordinator, makeOperationalJournaledWriteCoordinator } from "../write-coordination/journal/coordinator.ts";
 export { makeLocalLockRegistry } from "../persistence/local/local-lock-registry.ts";
 export { makeLocalVersionControlSystem } from "../persistence/git/local-version-control-system.ts";
 export { makeLocalAuthorityAttributionEventV2Log } from "../write-coordination/attribution/authority-attribution-event-v2-log.ts";
