+++ packages/kernel/src/store/daemon-runtime.ts
@@ -31,9 +31,9 @@ import {
 import type { DaemonDrainOptions, DaemonMaterializerBatchOptions } from "./daemon-runtime-options.ts";
 export type { DaemonDrainOptions, DaemonMaterializerBatchOptions } from "./daemon-runtime-options.ts";
 import { acquireDaemonGlobalLock, assertDaemonGlobalLockHeld, type DaemonGlobalLock } from "../write-coordination/journal/locks.ts";
-import { makeJournaledWriteCoordinator, makeOperationalJournaledWriteCoordinator, recoverJournaledWrites } from "./write-journal-coordinator.ts";
+import { makeJournaledWriteCoordinator, makeOperationalJournaledWriteCoordinator, recoverJournaledWrites } from "../write-coordination/journal/coordinator.ts";
 import type { OperationalActor } from "../write-coordination/journal/types.ts";
-import { writeOpTouchedPaths } from "./write-journal-operations.ts";
+import { writeOpTouchedPaths } from "../write-coordination/journal/operations/transaction-plan.ts";
 import {
   createDaemonProjectionGenerationManager,
   type DaemonProjectionGenerationManager,
