+++ packages/kernel/test/store/daemon-materializer-runtime.test.ts
@@ -8,7 +8,7 @@ import { createHarnessRuntimeContext, resolveHarnessLayout } from "../../src/lay
 import type { ProjectionSourceFenceFactory } from "../../src/ports/projection-source-fence.ts";
 import { projectionDatabaseSignature } from "../../src/projection/projection-generation-readiness.ts";
 import { rebuildTaskProjection } from "../../src/projection/sqlite-task-projection.ts";
-import { makeJournaledWriteCoordinator } from "../../src/store/write-journal-coordinator.ts";
+import { makeJournaledWriteCoordinator } from "../../src/write-coordination/journal/coordinator.ts";
 import { createDaemonRuntime, createMultiRepoDaemonRuntime } from "../../../adapters/local/src/index.ts";
 import { docWrite, withTempStoreAsync } from "./helpers.ts";
 import {
