+++ packages/kernel/test/store/entity-registry-substrate.test.ts
@@ -29,7 +29,7 @@ import {
 } from "../../src/entity/registry-compiler.ts";
 import { stablePayloadHash } from "../../src/integrity/stable-hash.ts";
 import { writeContentAddressedBlob } from "../../src/persistence/blob/content-addressed-blob-store.ts";
-import { makeJournaledWriteCoordinator } from "../../src/store/write-journal-coordinator.ts";
+import { makeJournaledWriteCoordinator } from "../../src/write-coordination/journal/coordinator.ts";
 import { withTempStore } from "./helpers.ts";
 
 const FixtureSchema = Schema.Struct({
